+++ packages/plugin-commons/web/src/index.ts
@@ -32,7 +32,6 @@ export { default as baseToolbarButton } from './Base/baseToolbarButton';
 
 // Components
 export { default as Image } from './Components/Image';
-export { default as Panel } from './Components/Panel';
 export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizontal';
 export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
 export { default as SelectionList } from './Components/SelectionList';