+++ packages/plugin-commons/web/src/Modals/UrlInputModal.js
@@ -6,7 +6,7 @@ import { KEYS_CHARCODE } from 'wix-rich-content-editor-common';
 import SettingsMobileHeader from './SettingsMobileHeader';
 import { CloseIcon } from '../Icons';
 import SettingsPanelFooter from '../Components/SettingsPanelFooter';
-import TextInput from '../Components/TextInput';
+import { TextInput } from 'wix-rich-content-ui-components';
 import { FOOTER_BUTTON_ALIGNMENT, MODAL_CONTROLS_POSITION } from '../consts';
 import styles from '../../statics/styles/url-input-modal.scss';