+++ packages/ui-components/web/src/index.ts
@@ -8,7 +8,7 @@ export { default as Button } from './Components/Button';
 // export { default as Image } from './Components/Image';
 export { default as InputWithLabel } from './Components/InputWithLabel';
 export { default as LabeledToggle } from './Components/LabeledToggle';
-// export { default as Panel } from './Components/Panel';
+export { default as Panel } from './Components/Panel';
 // export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizontal';
 // export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
 // export { default as SelectionList } from './Components/SelectionList';