+++ packages/plugin-commons/web/src/index.ts
@@ -37,11 +37,9 @@ export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizont
 export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
 export { default as SelectionList } from './Components/SelectionList';
 export { default as SettingsPanelFooter } from './Components/SettingsPanelFooter';
-export { default as SettingsSection } from './Components/SettingsSection';
 export { default as Slider } from './Components/Slider';
 export { default as SliderWithInput } from './Components/SliderWithInput';
 export { Tab, Tabs } from './Components/Tabs';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
-export { default as Loader } from './Components/Loader';
 
 export { getSelectionStyles } from './Utils/inlineStyleUtils';