+++ examples/main/src/Components/ExternalToolbar/ExternalToolbar.tsx
@@ -1,5 +1,5 @@
 import React, { Component, RefObject } from 'react';
-import { FileInput } from 'wix-rich-content-plugin-commons';
+import { FileInput } from 'wix-rich-content-ui-components';
 import { BUTTON_TYPES } from 'wix-rich-content-editor-common';
 import FormattingGroupButton from 'wix-rich-content-editor-common/libs/FormattingGroupButton';
 import FormattingDropdownButton from 'wix-rich-content-editor-common/libs/FormattingDropdownButton';