+++ packages/plugin-gallery/web/src/toolbar/inline-buttons.tsx
@@ -8,6 +8,7 @@ import {
   TranslationFunction,
   AnchorTarget,
   RelValue,
+  AvailableExperiments,
 } from 'wix-rich-content-common';
 import { GalleryPluginEditorConfig, GALLERY_TYPE } from '../types';
 
@@ -16,25 +17,29 @@ const createInlineButtons: CreateInlineButtons = ({
   anchorTarget,
   relValue,
   settings,
+  experiments = {},
   isMobile,
 }: {
   t: TranslationFunction;
   settings: GalleryPluginEditorConfig;
   anchorTarget: AnchorTarget;
   relValue: RelValue;
+  experiments: AvailableExperiments;
   isMobile: boolean;
 }) => {
   const modalStyles = getModalStyles({ isMobile });
   const icons = settings?.toolbar?.icons || {};
-  const spoilerButton = settings.spoiler
-    ? [
-        {
-          keyName: 'spoiler',
-          type: BUTTONS.SPOILER,
-          mobile: true,
-        },
-      ]
-    : [];
+  const { spoilerInInlineToolbar } = experiments;
+  const spoilerButton =
+    settings.spoiler && spoilerInInlineToolbar?.enabled
+      ? [
+          {
+            keyName: 'spoiler',
+            type: BUTTONS.SPOILER,
+            mobile: true,
+          },
+        ]
+      : [];
 
   return [
     {