+++ packages/plugin-image/web/src/consts.ts
@@ -10,3 +10,4 @@ export const DEFAULTS = Object.freeze({
 
 export const SMALL_SIZE_WIDTH = 350;
 export const SEO_IMAGE_WIDTH = 1000;
+export const DIVIDER = 'divider';