+++ CHANGELOG.md
@@ -16,6 +16,9 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `ricos-content`
+  - [#2594](https://github.com/wix-incubator/rich-content/pull/2594) list style removal
 
 </details>
 <hr/>
@@ -24,6 +27,7 @@
 ### :house: Internal
 - `gallery`
   - [#2590](https://github.com/wix/ricos/pull/2590) bump `ProGallery` to v3.1.14
+
 ## 8.42.2 (Jun 6, 2021)
 ### :bug: Bug Fix
 - `common`