+++ CHANGELOG.md
@@ -21,8 +21,7 @@
 
 - `ui-components`
   - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
-- `ui-components`
-  - [##2596](https://github.com/wix-incubator/rich-content/pull/#2596) moved fileInput component into ui-components folder
+
 
 </details>
 <hr/>