+++ CHANGELOG.md
@@ -17,6 +17,12 @@
     Click to see more.
   </summary>
 
+### :house: Internal
+
+- `ui-components`
+  - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
+
+
 </details>
 <hr/>