+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 - `ui-components`
   - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
 - `ui-components`
-  - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved RadioGroupHorizontal component into ui-components folder
+  - [#2598](https://github.com/wix-incubator/rich-content/pull/2598) moved RadioGroupHorizontal component into ui-components folder
 
 
 </details>