+++ CHANGELOG.md
@@ -20,8 +20,10 @@
 ### :house: Internal
 
 - `ui-components`
-  - [##2596](https://github.com/wix-incubator/rich-content/pull/#2596) moved fileInput component into ui-components folder
-  
+  - [#2596](https://github.com/wix-incubator/rich-content/pull/2596) moved fileInput component into ui-components folder
+- `ui-components` 
+  - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
+
 </details>
 <hr/>