+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `ui-components`
   - [#2577](https://github.com/wix-incubator/rich-content/pull/2577) moved Loader component into ui-components folder
+- `ui-components`
+  - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
 
 </details>
 <hr/>