+++ CHANGELOG.md
@@ -16,6 +16,15 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+  
+### :house: Internal
+
+- `ui-components`
+  - [#2577](https://github.com/wix-incubator/rich-content/pull/2577) moved Loader component into ui-components folder
+- `ui-components`
+  - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
+- `ui-components`
+  - [#2571](https://github.com/wix-incubator/rich-content/pull/2571) moved SettingsSection component into ui-components folder
 
 </details>
 <hr/>