+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 ### :house: Internal
 
 - `ui-components`
+  - [#2596](https://github.com/wix-incubator/rich-content/pull/2596) moved fileInput component into ui-components folder
+  - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
   - [#2577](https://github.com/wix-incubator/rich-content/pull/2577) moved Loader component into ui-components folder
   - [#2582](https://github.com/wix-incubator/rich-content/pull/2582) moved Panel component into ui-components folder
   - [#2604](https://github.com/wix-incubator/rich-content/pull/2604) moved TextInput component into ui-components folder