+++ CHANGELOG.md
@@ -27,6 +27,11 @@
   - [#2604](https://github.com/wix-incubator/rich-content/pull/2604) moved TextInput component into ui-components folder
   - [#2571](https://github.com/wix-incubator/rich-content/pull/2571) moved SettingsSection component into ui-components folder
 
+### :rocket: New Feature
+
+- `spoiler`
+  - [#2326](https://github.com/wix/ricos/pull/2326) remove spoiler from inline toolbar and add it to settings modal
+
 </details>
 <hr/>